refactor(AppData): Move AppData code back to MyLayout.
