+++ pages/_app.js
@@ -2,7 +2,6 @@ import App from 'next/app'
 import Router from 'next/router'
 import { Auth0Provider } from 'use-auth0-hooks'
 
-import AppData from '../components/AppData'
 import Layout from '../components/MyLayout'
 import { getAuthRedirectUri } from '../util/auth'
 import { AUTH0_SCOPE } from '../util/constants'
@@ -79,9 +78,7 @@ export default class Root extends App {
         onRedirectCallback={onRedirectCallback}
       >
         <Layout>
-          <AppData>
-            <Component {...pageProps} />
-          </AppData>
+          <Component {...pageProps} />
         </Layout>
       </Auth0Provider>
     )
